requirements.txt
numpy
matplotlib
tensorflow
Pillow

